Create Docker image for development on Ubuntu (#1)
* Add Dockerfile, configure script, and Docker helper scripts

* Fix Dockerfile and configure script

* Update GitHub Actions workflow to build Docker image

* Make configure.sh script executable

* Update configure.sh script and Dockerfile

* Add thirdparty dependencies as submodules

* Fix thirdparty library linking in CMakeList

* Fix build errors and warnings

* Update CPack for Debian packages

+++ src/ECS/include/Mesh.h
@@ -16,7 +16,7 @@
 
 typedef struct {
     float *data;
-    unsigned int size;
+    size_t size;
     unsigned int stride;
 } Buffer;
